fix(Sortir): fix some bug

+++ Modules/Transaction/Resources/views/sortir/show.blade.php
@@ -160,7 +160,7 @@
             const inputs = document.querySelectorAll('.quantity-input');
             const stockEl = document.querySelector('.stock');
 
-            const initialStock = parseInt(stockEl.textContent.trim()) || 0;
+            const initialStock = parseFloat(stockEl.textContent.trim()) || 0;
 
             function updateStock() {
                 let totalQty = 0;
